SYAI/문서 자동생성/requirements.txt
google-generativeai>=0.7.0
pypdf>=4.2.0
pdfplumber>=0.11.0
python-docx>=1.1.0
markdown>=3.6
scikit-learn>=1.4.2
tqdm>=4.66.0
orjson>=3.10.0
rapidfuzz>=3.9.3
